+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## <small>0.0.1-0-monorepo (2018-12-16)</small>
 
-* feat(frontend, gateway, services, core): Add media service and add withMedia parameter to context se ([5fddbe5](https://gitlab.com/pojntfx/learn-chinese-platform/commit/5fddbe5))
+* feat(frontend): Merge media and definitions ([9dc1952](https://gitlab.com/pojntfx/learn-chinese-platform/commit/9dc1952))
 * feat(frontend): Show stroke animation ([9b1c71b](https://gitlab.com/pojntfx/learn-chinese-platform/commit/9b1c71b))
+* feat(frontend, gateway, services, core): Add media service and add withMedia parameter to context se ([5fddbe5](https://gitlab.com/pojntfx/learn-chinese-platform/commit/5fddbe5))